Cleanup interfaces in utils.
