+++ src/utils/index.ts
@@ -3,11 +3,6 @@ import { base32 } from 'multiformats/bases/base32'
 import { peerIdFromString } from '@libp2p/peer-id'
 import type { PeerId } from '@libp2p/interface-peer-id'
 
-import type { IdentityModule } from '@/identity/interface.js'
-import type { AccessModule } from '@/access/interface.js'
-import type { StoreModule } from '@/store/interface.js'
-import type { EntryModule } from '@/entry/interface.js'
-
 export const cidstring = (cid: CID | string): string => cid.toString(base32)
 export const parsedcid = (string: string): CID => CID.parse(string, base32)
 
@@ -17,14 +12,3 @@ export const decodedcid = (bytes: Uint8Array): CID => CID.decode(bytes)
 export const peerIdString = (peerId: PeerId): string =>
   peerId.toCID().toString(base32)
 export const parsedPeerId = (peerId: string): PeerId => peerIdFromString(peerId)
-
-export interface DirsReturn {
-  [name: string]: string
-}
-
-export interface Components {
-  access: AccessModule
-  entry: EntryModule
-  identity: IdentityModule
-  store: StoreModule
-}
